+++ src/Controllers/SCStormRunoffController.ts
@@ -1071,10 +1071,6 @@ module StreamStats.Controllers {
             this.canContinue = true;
             this.SelectedAEP = {"name": "50%", "value": 50};
             this.SelectedAEPSynthetic = {"name": "10%", "value": 10};
-            this.SelectedStandardCurve = { "name": "Area Weighted CN", "value": 1 }
-            this.SelectedCNModification = { "name": "McCuen", "value": 1 }
-            //this.SelectedTimeOfConcentration = { "name": "Travel Time Method", "value": 1 }
-            this.SelectedRainfallDistribution = { "name": "Type II", "value": 2 }
 
             this._chosenFlowTypeIndex = null;
         }