+++ src/Controllers/ReportController.js
@@ -138,6 +138,21 @@ var StreamStats;
                 enumerable: false,
                 configurable: true
             });
+            ReportController.prototype.setStream = function (stream) {
+                this.studyAreaService.selectedStudyArea.NHDStream = stream;
+                var input = document.getElementById(stream.GNIS_NAME);
+                if (input != null) {
+                    input.checked = true;
+                }
+                else {
+                    setTimeout(function () {
+                        var input = document.getElementById(stream.GNIS_NAME);
+                        if (input != null) {
+                            input.checked = true;
+                        }
+                    }, 1000);
+                }
+            };
             ReportController.prototype.selectFDCTMTab = function (tabname) {
                 if (this.selectedFDCTMTabName == tabname)
                     return;