+++ src/appConfig.js
@@ -13,8 +13,7 @@ configuration.baseurls =
         'ScienceBase': 'https://gis.usgs.gov/sciencebase2',
         'GageStatsServices': 'https://test.streamstats.usgs.gov/gagestatsservices',
         'WeightingServices': 'https://streamstats.usgs.gov/channelweightingservices',        
-        'FlowAnywhereRegressionServices': 'https://streamstats.usgs.gov/regressionservices',
-        'FlowAnywhereMapServices': 'https://gis.wim.usgs.gov',
+        'FlowAnywhereRegressionServices': 'https://streamstats.usgs.gov/regressionservices'
     };
 
 //override streamstats arguments if on production, these get overriden again in MapController after load balancer assigns a server