+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   - Flow Anywhere Method from StreamEst
 
   ### Changed
+  - Channel Weighting services URL 
 
   - NWIS Page link from legacy real-time page to Next Generation Monitoring Location page