+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
     -  South Carolina application selects nearest NHD stream and HUC 8 watershed and adds to report
     -  South Carolina application allows user to select a different NHD stream via radio buttons in report
     -  Query NHD streams in South Carolina and add new section to report
+    -  Added Synthetic Unit Hydrograph functionality to the SC Storm Runoff model
+
 
   ### Changed