+++ CHANGELOG.md
@@ -11,8 +11,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
   ### Changed
   - Channel Weighting services URL 
-
   - NWIS Page link from legacy real-time page to Next Generation Monitoring Location page
+  - FlowAnywhereMapServices to StreamStatsMapServices because IowaStreamEst map services were moved from gis.wim.usgs.gov to gis.streamstats.usgs.gov
 
   ### Fixed